Implemented DUMMY GUI
GUI now works and can go from frame to frame. Lots of logic and checking required, but there are now action listeners between several of the screens for the happy-path of the project. Currently no logic is setup to store or read cards, it just prints inputs to the console.